Changes to documentation and directory structure. * Modified docs as suggested on the PR. * Fixed formatting for a few pages. * Renamed common directory to user as it will contain user specific code. * Updated user management command.
