+++ edx_django_utils/management/README.rst
@@ -1,5 +1,6 @@
-# Django Management Commands for Multiple Services
+Django Management Commands for Multiple Services
+===============================
 
-Management commands useful to multiple django apps will live here.
+Management commands useful to multiple django services will live here.
 
 .. _User and Group Management Commands: /docs/decisions/0005-user-and-group-management-commands.rst
