+++ edx_django_utils/management/tests/__init__.py
@@ -1,1 +1,1 @@
-"""Tests for Student Management Commands."""
+"""Tests for Management Commands."""
